Suggested rewrite of 'poor quality' paragraph
I think the extra information here slightly weakens your point. We
would (should) do TDD and code reviews even if we were a fully Ruby
shop, so they are really just part of the bigger picture of how we
attempt to maintain quality. Pair programming is relevant to your
point because it's how we deal with the explicit fact of people using
tools they are not familar with.

Also, you've linked to an article about TDD, no need to spell it out
here.

Broadly, I've made two changes here - one is to focus on the way pair
programming mitigates the problem of unfamiliarity with the tools, and
the other is to cut back the description of TDD. Please feel free to
entirely ignore both.